feat(macos): implement "open firewall settings" button
