+++ app/lib/util/native/macos_channel.dart
@@ -54,6 +54,10 @@ Future<bool> isReduceMotionEnabledMacOs() async {
   return await _methodChannel.invokeMethod('isReduceMotionEnabled') ?? false;
 }
 
+Future<void> openFirewallSettings() async {
+  await _methodChannel.invokeMethod('openFirewallSettings');
+}
+
 // This happens:
 /// - on macOS when text is dropped onto the app Dock icon
 /// - on macOS when text is dropped onto the app menu bar icon
